Create subfolder for combat-related utils
