+++ Utilities/Combat/Action.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Action(Enum):
+    ATTACK = 1
+    BLOCK = 2
+    PARRY = 3
+    HEAL = 4
+    BIDE = 5
